broken fileIO code
main is just a simple test code to read in a couple of lines from test.txt and cout them to the screen. the other two files are my failed attempt to use JP's class implementation. in my opinion, the class just seems to make the whole thing overly complicated, but it is probably just the fact that I am trying to follow someone else's instructions rather than doing it the way that would be intuitive to me. 

JP, can you take a look at the code and see why it is throwing fatal errors? if i comment out lines 15-20 in main.cpp it compiles and runs fine.